InnerNonbondedForce moved to cvlib
