+++ docs/zref.bib
@@ -391,3 +391,16 @@
 	title = {Unified Efficient Thermostat Scheme for the Canonical Ensemble with Holonomic or Isokinetic Constraints via Molecular Dynamics},
 	journal = {The Journal of Physical Chemistry A}
 }
+
+@article{Morrone_2010,
+	doi = {10.1021/ct100054k},
+	year = 2010,
+	month = {may},
+	publisher = {American Chemical Society ({ACS})},
+	volume = {6},
+	number = {6},
+	pages = {1798--1804},
+	author = {Joseph A. Morrone and Ruhong Zhou and B. J. Berne},
+	title = {Molecular Dynamics with Multiple Time Scales: How to Avoid Pitfalls},
+	journal = {Journal of Chemical Theory and Computation}
+}
